requirements.txt
fastapi
uvicorn
rank_bm25
psycopg2-binary
transformers
sentence-transformers
qdrant-client
wikipedia-api
datasets
